Make checkbox labels clickable
Been using this plugin for a while, it's awesome.

Just noticed this today - would be nice if I could click the label to set the checkbox - just gives a bigger area to click. I've been working on a project where I need to manually create cookies to test things and this would just help speed that up!